requirements.txt
flask==3.0.0
pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
piexif==1.1.3
tensorflow==2.15.0
